# wallet: Reset reused transactions cache