@kallewoof 2020-01-03
Github spazzing out, but https://github.com/bitcoin/bitcoin/pull/17843#discussion_r362716618 is in response to @instagibbs suggestion to `break`. I think breaking would skip over any remaining UTXO:s sent to the same address.