@kallewoof 2020-01-03
@promag Would love to see worst case scenario analysis (is this O(n) or O(n^2) or worse?).