@kallewoof 2020-01-03
> @kallewoof I think it's O(n). Problem is that for big wallets, all transaction outputs will be scanned as much as the input count of the new transaction.

So if `n` is wallet size and `m` is input count, this would be `O(nm)`. That's pretty bad, but honestly don't think it's a huge deal. I could be wrong, though.