@instagibbs 2020-01-03
This might make a good topic for IRC meeting. I think making this slower is ok provided it's opt-in, which it appears to be based on the required avoid_reuse wallet flag guarding `SetUsedDestinationState`. An algorithmic cleanup is of course encouraged.