@achow101 2020-01-03
I agree with @promag. I think we should use `std::unordered_multimap` and accumulate the destinations so that it can all be done with one iteration over the fewest `CWalletTx`s needed. `O(nm)` can be pretty slow. We already know that for large wallets, things that do this kind of iteration can take a while so I think we should try to avoid further iteration like this.