@promag 2020-01-03
I also think it's fine to improve in a follow up and have the fix merged. In that case I'd just change this to do as suggested:
> accumulate all CTxDestination to be marked dirty and then iterate the wallet transactions just once.
which is pretty simple.