@fjahr 2020-01-06
Pushed update including most of the feedback and implementing the simpler solution suggestion by @promag:

> * accumulate all `CTxDestination` to be marked dirty and then iterate the wallet transactions just once.