@fjahr 2020-01-08
Rebased and addressed feedback.