@achow101 2020-01-08
ACK 7618ea2cc02042cd20c633de9d29c34b5a1179b0

Just suggested changes and rebase since last review.