@promag 2020-01-08
I think you could revert latest change regarding amount caches considering what's done in #17889.