@fjahr 2020-01-08
> I think you could revert latest change regarding amount caches considering what's done in #17889.

Removed skipping dirty tx again to make it a more coherent change together with #17889. Sorry @achow101 .