@kallewoof 2020-01-11
I did some rough benchmarks to see how this would affect biggish wallets, and am not seeing any significant drops in performance:

```bash
# 2 btc per send

# Master:
#
# UTXO count    Total Time      Transactions    Time per transaction
# 100           00:57           160             0.356
# 500           00:56           160             0.350
# 1000          01:09           160             0.431
# 2000          01:11           160             0.444
# 5000          01:55           160             0.719

# With patch:
#
# UTXO count    Total Time      Transactions    Time per transaction
# 100           01:05           160             0.406
# 500           00:56           160             0.350
# 1000          01:00           160             0.375
# 2000          01:22           160             0.513
# 5000          02:14           160             0.838

# 1/20th of balance per send, gen per 10 tx

# Master:
#
# UTXO count    Total Time      Transactions    Time per transaction
# 100           01:05           160             0.406
# 500           01:00           160             0.375
# 1000          01:06           160             0.800
# 2000          01:03           160             0.394
# 5000          01:21           160             0.900

# With patch:
#
# UTXO count    Total Time      Transactions    Time per transaction
# 100           01:01           160             0.381
# 500           00:59           160             0.369
# 1000          01:02           160             0.388
# 2000          01:06           160             0.800
# 5000          01:40           160             0.625
```