@fjahr 2020-01-13
> I did some rough benchmarks to see how this would affect biggish wallets, and am not seeing any significant drops in performance:

Thanks a lot for running these!

Pushed update to outdated comment.