@fanquake 2020-02-06
Being backported in #18083.